web-app/backend/requirements.txt
fastapi
uvicorn[standard]
imfusion-sdk==0.10.0
numpy
Pillow
python-dotenv
